Update index.html
I completely refactored this page to version 4. I am happy with the design as the page is not to big and not to small. I found the carousels good but they were to big for mobile. I would like to hit both small mobile and desktop viewers. Deliverance of project materiel is viewable by code only the github. I will work on linking active program materiels again in the future.